Improved matlab test script - handles faulty test removal

+++ scripts/matlab/removeFaultyTests.m
@@ -1,8 +1,11 @@
-function [ ] = removeFaultyTests(test_path, possible_failed_mistakes, possible_success_mistakes )
+function [ removed_tests ] = removeFaultyTests(test_path, possible_failed_mistakes, possible_success_mistakes )
 %UNTITLED Summary of this function goes here
 %   Detailed explanation goes here
 
 % test_path = '~/counter_uas/scripts/matlab/skyler/recent_tests/analysis/autotest/test16/test/';
+
+removed_tests = [];
+
 
 % Handle failed tests
 for i=1:size(possible_failed_mistakes,2)
@@ -14,7 +17,10 @@ for i=1:size(possible_failed_mistakes,2)
         [T, data] = evalc('processAllTopics(bagfile);');
     catch
         disp(['Removing ' bagfile]);
-        system(['rm ' bagfile ' ' results_file]);
+        if(exist(bagfile,'file'))
+            system(['rm ' bagfile ' ' results_file]);
+            removed_tests = [removed_tests, test_number];
+        end        
         continue;
     end
     
@@ -30,9 +36,14 @@ for i=1:size(possible_failed_mistakes,2)
     %  vvv Check for UAV crash   vvv Check for no data   vvv Check for intruder crash
     if(range(x_ranges) > 30 || size(uav1_pose,2) == 0 || range(intruder_pose(2,:)) == 0)
        disp(['Removing ' bagfile]);
-       system(['rm ' bagfile ' ' results_file]);
+        if(exist(bagfile,'file'))
+            system(['rm ' bagfile ' ' results_file]);
+            removed_tests = [removed_tests, test_number];
+        end 
     else
-       disp(['NOT removing ' bagfile]);
+        if(exist(bagfile,'file'))        
+            disp(['NOT removing ' bagfile]);
+        end
     end
 
 end
@@ -47,7 +58,10 @@ for i=1:size(possible_success_mistakes,2)
         [T, data] = evalc('processAllTopics(bagfile);');
     catch
         disp(['Removing ' bagfile]);
-        system(['rm ' bagfile ' ' results_file]);
+        if(exist(bagfile,'file'))
+            system(['rm ' bagfile ' ' results_file]);
+            removed_tests = [removed_tests, test_number];
+        end
         continue;
     end
     
@@ -63,9 +77,16 @@ for i=1:size(possible_success_mistakes,2)
     %  vvv Check for UAV crash   vvv Check for no data   vvv Check for intruder crash
     if(range(x_ranges) > 30 || size(uav1_pose,2) == 0 || range(intruder_pose(2,:)) == 0)
        disp(['Removing ' bagfile]);
-       system(['rm ' bagfile ' ' results_file]);
+        if(exist(bagfile,'file'))
+            system(['rm ' bagfile ' ' results_file]);
+            removed_tests = [removed_tests, test_number];
+        end
     else
-       disp(['NOT removing ' bagfile]);
+        if(exist(bagfile,'file'))        
+            disp(['NOT removing ' bagfile]);
+        end
     end
+    
+    removed_tests = sort(removed_tests);
 end
